fix(diagrams): make `EXPORT_FOLDER` logic more robust
Before these changes when I tried to export diagram artifacts to
`./gen` nothing was written to the expected directory.

    $ export EXPORT_FOLDER="./gen"
    $ cargo run --example traffic_light --all-features

I discovered the files were actually written, but to unexpected
locations.

These changes eliminate assumptions about how the path is specified
and also attempt to create the `EXPORT_FOLDER` if it doesn't already
exist.